Refact  message functions to  controller
