+++ views/chatRoom.ejs
@@ -5,6 +5,10 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>ChatJS</title>
         <link rel="stylesheet" href="./assets/style.css" />
+        <script type="module">
+            import {io} from "https://cdn.socket.oi/4.3.2/socket.io.esm.min.js"
+
+        </script>
     </head>
     <body>
         <section>
